Move provider log code to file
